% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_version.R
\name{compare_version}
\alias{compare_version}
\title{compare_version}
\usage{
compare_version(package, version)
}
\arguments{
\item{package}{The package name}

\item{version}{The version to compare it against}
}
\description{
compare_version
}
