% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_environments.r
\docType{data}
\name{config}
\alias{config}
\title{Environment to store config
test}
\format{
An object of class \code{environment} of length 0.
}
\usage{
config
}
\description{
Environment to store config
test
}
\keyword{datasets}
