% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_Documentation.R
\docType{data}
\name{DailyCatch}
\alias{DailyCatch}
\title{Daily fish catch data}
\format{
A dataframe with 5065 rows and 8 variables:
\describe{
  \item{CatchRecord}{Unique value for each daily catch}
  \item{WeightCaught}{Total weight in kg for daily catch}
  \item{Station}{Broad location descriptor of daily catch e.g. Community fishery}
  \item{Latitude}{Latitude of fishing location}
  \item{Longitude}{Longitude of fishing location}
  \item{Habitat}{Habitat of fishing location}
  \item{Location}{Local location descriptor of daily catch}
 ...
}
}
\source{
Conservation International - Greater Mekong
}
\usage{
DailyCatch
}
\description{
A dataset containing daily fish catch data from the Tonle Sap Lake
in Cambodia
}
\keyword{datasets}
