% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardEmailSpecific}
\alias{DashboardEmailSpecific}
\title{Sends out mass emails that are stored in an xlsx file}
\usage{
DashboardEmailSpecific(emailBCC, emailSubject, emailText, emailFooter = TRUE,
  OAUTHLocation = PROJ$DEFAULT_EMAILS_OAUTH_LOCATION)
}
\arguments{
\item{emailBCC}{a}

\item{emailSubject}{a}

\item{emailText}{a}

\item{emailFooter}{a}

\item{OAUTHLocation}{a}
}
\description{
Sends out mass emails that are stored in an xlsx file
}
