% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardFolder}
\alias{DashboardFolder}
\title{Dashboard folders}
\usage{
DashboardFolder(inside = "data_raw", f = NULL)
}
\arguments{
\item{inside}{where it is inside}

\item{f}{an optional file}
}
\description{
This function relies upon you using the dashboard folder setup: \code{vignette("dashboardbasics", package = "fhi")}.
This function finds folders according to the dashboard philosophy
}
