% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardInitialise}
\alias{DashboardInitialise}
\title{If folders are setup according to the
dashboard philosophy, then this function
sets RPROJ}
\usage{
DashboardInitialise(STUB = "/", SRC = "src", NAME = NULL,
  changeWorkingDirToTmp = TRUE)
}
\arguments{
\item{STUB}{a}

\item{SRC}{a}

\item{NAME}{a}

\item{changeWorkingDirToTmp}{a}
}
\description{
If folders are setup according to the
dashboard philosophy, then this function
sets RPROJ
}
