% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardIsDev}
\alias{DashboardIsDev}
\title{Is the dashboard in development?}
\usage{
DashboardIsDev()
}
\description{
This function relies upon you using the dashboard folder setup: \code{vignette("dashboardbasics", package = "fhi")}
}
