% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardIsInitialised}
\alias{DashboardIsInitialised}
\title{Is the dashboard initialised?}
\usage{
DashboardIsInitialised()
}
\description{
This function relies upon you using the dashboard folder setup: \code{vignette("dashboardbasics", package = "fhi")}
}
