% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardIsProduction}
\alias{DashboardIsProduction}
\title{Is the dashboard in production?}
\usage{
DashboardIsProduction()
}
\description{
This function relies upon you using the dashboard folder setup: \code{vignette("dashboardbasics", package = "fhi")}
}
