% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardMsg}
\alias{DashboardMsg}
\title{Messaging}
\usage{
DashboardMsg(txt, type = "msg", syscallsDepth = 2)
}
\arguments{
\item{txt}{Text}

\item{type}{msg, warn, err}

\item{syscallsDepth}{The number of syscalls included in the message. Set to 0 to disable.}
}
\description{
Messaging
}
