% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV_DvNF.R
\name{DvNF_gauge}
\alias{DvNF_gauge}
\title{calculate DvNF for measured flow/gauge height data}
\usage{
DvNF_gauge(x, y, z)
}
\arguments{
\item{x}{input file}

\item{y}{start year for the regulated period}

\item{z}{end year of the unregulated data}
}
\description{
calculate DvNF for measured flow/gauge height data
}
\examples{
DvNF_gauge(PK_HT, 2014, 1990)
}
