% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV_DvNF.R
\name{DvNF_model}
\alias{DvNF_model}
\title{calculate DvNF for modelled flow/gauge height data}
\usage{
DvNF_model(x)
}
\arguments{
\item{x}{input file}
}
\description{
calculate DvNF for modelled flow/gauge height data
}
\examples{
DvNF_model(KL_MOD)
}
