% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_DFC.R
\name{ES_DFC}
\alias{ES_DFC}
\title{Calculate the Ecosystem Service for daily fish catch using Method 2 and a single threshold this will give a single number across multiple CFi's.}
\usage{
ES_DFC(x, DFC_thresh, S_Date, E_Date)
}
\arguments{
\item{x}{input file}

\item{DFC_thresh}{Daily catch threshold weight}

\item{S_Date}{Start date for the assessment period}

\item{E_Date}{End date for the assessment period}
}
\description{
Calculate the Ecosystem Service for daily fish catch using Method 2 and a single threshold this will give a single number across multiple CFi's.
}
\examples{
ES_DFC(DailyCatch, 1.5, '2015-01-01', '2021-01-08')
}
