% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_DFC.R
\name{ES_DFC_S}
\alias{ES_DFC_S}
\title{Calculate the Ecosystem Service for daily fish catch using Method 2 and a single threshold this will give a single number for a single CFi.}
\usage{
ES_DFC_S(x, CFi, DFC_thresh, S_Date, E_Date)
}
\arguments{
\item{x}{input file}

\item{CFi}{Community fishery for which the metric will be calculated, this is listed under Station in the input file}

\item{DFC_thresh}{Daily catch threshold weight}

\item{S_Date}{Start date for assessment period}

\item{E_Date}{End date for assessment period}
}
\description{
Calculate the Ecosystem Service for daily fish catch using Method 2 and a single threshold this will give a single number for a single CFi.
}
\examples{
ES_DFC_S(DailyCatch, "Balat", 1.5, '2018-01-01', '2021-08-31')
}
