% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_DFC.R
\name{ES_DFC_SM}
\alias{ES_DFC_SM}
\title{Calculate the Ecosystem Service for daily fish catch using Method 2 and monthly thresholds for a single CFi \ Station}
\usage{
ES_DFC_SM(
  x,
  CFi,
  Jan_t,
  Feb_t,
  Mar_t,
  Apr_t,
  May_t,
  Jun_t,
  Jul_t,
  Aug_t,
  Sep_t,
  Oct_t,
  Nov_t,
  Dec_t,
  S_Date,
  E_Date
)
}
\arguments{
\item{x}{input file}

\item{CFi}{Community fishery for which the metric will be calculated, this is listed under Station in the input file}

\item{Jan_t}{Daily catch threshold weight for January}

\item{Feb_t}{Daily catch threshold weight for February}

\item{Mar_t}{Daily catch threshold weight for March}

\item{Apr_t}{Daily catch threshold weight for April}

\item{May_t}{Daily catch threshold weight for May}

\item{Jun_t}{Daily catch threshold weight for June}

\item{Jul_t}{Daily catch threshold weight for July}

\item{Aug_t}{Daily catch threshold weight for August}

\item{Sep_t}{Daily catch threshold weight for September}

\item{Oct_t}{Daily catch threshold weight for October}

\item{Nov_t}{Daily catch threshold weight for November}

\item{Dec_t}{Daily catch threshold weight for December}

\item{S_Date}{Start date for assessment period}

\item{E_Date}{End date for assessment period}
}
\description{
Calculate the Ecosystem Service for daily fish catch using Method 2 and monthly thresholds for a single CFi \ Station
}
\examples{
ES_DFC_SM(DailyCatch, "Balat", 50, 50, 20, 20, 10, 50, 50,
50, 50, 50, 50, 50, '2021-01-01', '2021-12-31')
}
