% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_Indicator.R
\name{ES_M1}
\alias{ES_M1}
\title{Calculate the Ecosystem Service Indicator using Method 1}
\usage{
ES_M1(x, Threshold)
}
\arguments{
\item{x}{input file}

\item{Threshold}{upper threshold value}
}
\description{
Calculate the Ecosystem Service Indicator using Method 1
}
\examples{
ES_M1(ES_test, 0.13)
}
