% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_Indicator.R
\name{ES_M2}
\alias{ES_M2}
\title{Calculate the Ecosystem Service Indicator using Method 2}
\usage{
ES_M2(x, Threshold)
}
\arguments{
\item{x}{input file}

\item{Threshold}{upper threshold value}
}
\description{
Calculate the Ecosystem Service Indicator using Method 2
}
\examples{
ES_M2(ES_test, 0.09)
}
