% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_Documentation.R
\docType{data}
\name{ES_test}
\alias{ES_test}
\title{Ecosystem Service data}
\format{
A dataframe with 395 rows and 2 variables:
\describe{
  \item{Name}{Site name}
  \item{Variable}{Measured Ecosystem service variable}
 ...
}
}
\source{
Mekong River Commission
}
\usage{
ES_test
}
\description{
A dataset containing dummy Ecosystem service data
}
\keyword{datasets}
