% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV_WQ.R
\name{EV_pH}
\alias{EV_pH}
\title{Calculate the Ecosystem Vitality Water Quality Indicator for pH using Method 2}
\usage{
EV_pH(x, pH_low, pH_high)
}
\arguments{
\item{x}{input file}

\item{pH_low}{lower threshold value}

\item{pH_high}{upper threshold value}
}
\description{
Calculate the Ecosystem Vitality Water Quality Indicator for pH using Method 2
}
\examples{
EV_pH(Wq3S_Cam, 6, 9)
}
