% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV_WQ.R
\name{EV_TN}
\alias{EV_TN}
\title{Calculate the Ecosystem Vitality Water Quality Indicator for Total Nitrogen using Method 2}
\usage{
EV_TN(x, TN_low, TN_high)
}
\arguments{
\item{x}{input file}

\item{TN_low}{lower threshold value}

\item{TN_high}{upper threshold value}
}
\description{
Calculate the Ecosystem Vitality Water Quality Indicator for Total Nitrogen using Method 2
}
\examples{
EV_TN(Wq3S_Cam, 0, 1.6)
}
