% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV_WQ.R
\name{EV_TP}
\alias{EV_TP}
\title{Calculate the Ecosystem Vitality Water Quality Indicator for Total Phosphorus using Method 2}
\usage{
EV_TP(x, TP_high)
}
\arguments{
\item{x}{input file}

\item{TP_high}{upper threshold value}
}
\description{
Calculate the Ecosystem Vitality Water Quality Indicator for Total Phosphorus using Method 2
}
\examples{
EV_TP(Wq3S_Cam, 0.13)
}
