% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV_WQ_TSS.R
\name{EV_TSS}
\alias{EV_TSS}
\title{Calculate the Ecosystem Vitality Water Quality Indicator for Total Suspended Solids where thresholds are unknown using Method 2}
\usage{
EV_TSS(x, y)
}
\arguments{
\item{x}{input file}

\item{y}{threshold year this is the year previous to the first assessment year}
}
\description{
Calculate the Ecosystem Vitality Water Quality Indicator for Total Suspended Solids where thresholds are unknown using Method 2
}
\examples{
EV_TSS(Wq3S_Cam, 2008)
}
