% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV_WQ_G.R
\name{EV_WQ}
\alias{EV_WQ}
\title{Calculate the Ecosystem Vitality Water Quality Indicator for a parameter within lower and upper bounds using Method 2}
\usage{
EV_WQ(x, parameter, WQ_low, WQ_high)
}
\arguments{
\item{x}{input file}

\item{parameter}{column name of water quality parameter to be assessed}

\item{WQ_low}{lower water quality threshold value}

\item{WQ_high}{upper water quality threshold value}
}
\description{
Calculate the Ecosystem Vitality Water Quality Indicator for a parameter within lower and upper bounds using Method 2
}
\examples{
EV_WQ(Wq3S_Cam, "TOTN_mgL", 0, 1.6)
}
