% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV_WQ_G.R
\name{EV_WQU}
\alias{EV_WQU}
\title{Calculate the Ecosystem Vitality Water Quality Indicator for parameters with undefined thresholds using Method 2}
\usage{
EV_WQU(x, parameter, cutoff)
}
\arguments{
\item{x}{input file}

\item{parameter}{column name of water quality parameter to be assessed}

\item{cutoff}{Threshold year. This is the final year of the pre assessment period, thus if your impact period starts in 2012, enter 2011}
}
\description{
Calculate the Ecosystem Vitality Water Quality Indicator for parameters with undefined thresholds using Method 2
}
\examples{
EV_WQU(Wq3S_Cam, "TOTP_mgL", 2008)
}
