% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_stable.R
\name{file_stable}
\alias{file_stable}
\title{Is the file stable (size not increasing/decreasing)?}
\usage{
file_stable(file = tempfile(), delay = 10)
}
\arguments{
\item{file}{The file of interest}

\item{delay}{Number of seconds between checking file size a second time}
}
\description{
Is the file stable (size not increasing/decreasing)?
}
