% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.r
\name{FortifyShapeFile}
\alias{FortifyShapeFile}
\title{Reads in shapefile and fortifies it}
\usage{
FortifyShapeFile(file, region)
}
\arguments{
\item{file}{Shapefile location}

\item{region}{Name of the variable to fortify on}
}
\description{
Reads in shapefile and fortifies it
}
