% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{GenData}
\alias{GenData}
\title{Creates the NorwayLocations, NorwayMunicipMerging, and NorwayPopulation data.table}
\usage{
GenData(saveLoc = file.path("inst", "createddata"))
}
\arguments{
\item{saveLoc}{Location of data}
}
\description{
Creates the NorwayLocations, NorwayMunicipMerging, and NorwayPopulation data.table
}
