% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{GenNorwayLocations}
\alias{GenNorwayLocations}
\title{Creates the NorwayLocations data.table}
\usage{
GenNorwayLocations(saveLoc = file.path("inst", "createddata"))
}
\arguments{
\item{saveLoc}{Location to save file to}
}
\description{
Creates the NorwayLocations data.table
}
