% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{GenNorwayMunicipMerging}
\alias{GenNorwayMunicipMerging}
\title{Creates the Norway Municip Merging (kommunesammenslaping) data.table}
\usage{
GenNorwayMunicipMerging(saveLoc = file.path("inst", "createddata"))
}
\arguments{
\item{saveLoc}{Location to save file to}
}
\description{
Last updated 2017-07-29
}
