% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_location_utilities.R
\name{get_location_name}
\alias{get_location_name}
\title{location utilities}
\usage{
get_location_name(location_code)
}
\arguments{
\item{location_code}{a location code}
}
\description{
location utilities
}
