% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{\%dopardev\%}
\alias{\%dopardev\%}
\title{\%dopardev\%}
\usage{
obj \%dopardev\% ex
}
\arguments{
\item{obj}{foreach object used to control the evaluation of ex.}

\item{ex}{the R expression to evaluate.}
}
\description{
\%dopardev\%
}
