% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoyearweek.R
\name{isoyearweek}
\alias{isoyearweek}
\title{`isoyearweek` vector}
\usage{
isoyearweek(x = integer())
}
\arguments{
\item{x}{A numeric vector}
}
\value{
An S3 vector of class `fhi_isoyearweek`.
}
\description{
This creates a double vector that represents percentages so when it is
printed, it is multiplied by 100 and suffixed with `%`.
}
\examples{
isoyearweek("2020-01")
}
