% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_Documentation.R
\docType{data}
\name{KL_MOD}
\alias{KL_MOD}
\title{Modelled flow/guage height data}
\format{
A dataframe with 8766 rows and 3 variables:
\describe{
  \item{Date}{Date of the format, YYYY/MM/DD}
  \item{Regulated}{Modelled regulated flow/guage height}
  \item{Unregulated}{Modelled unregulated flow/guage height}
 ...
}
}
\source{
Mekong River Commission
}
\usage{
KL_MOD
}
\description{
A dataset containing the modelled height of the Tonle Sap Lake at Kampong Luong from 1 January 1985
to 31 December 2008
}
\keyword{datasets}
