% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{Log}
\alias{Log}
\title{Logs values to LOGDATA$log}
\usage{
Log(name, val = Sys.time(), log = "log1")
}
\arguments{
\item{name}{Name of value being logged}

\item{val}{Value being logged}

\item{log}{Name of log}
}
\description{
Logs values to LOGDATA$log
}
