% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{make_skeleton}
\alias{make_skeleton}
\title{make_skeleton}
\usage{
make_skeleton(
  date_min = NULL,
  date_max = NULL,
  yrwk_min = NULL,
  yrwk_max = NULL,
  location_code = NULL,
  granularity_geo = "all",
  location_reference = fhidata::norway_locations_long_b2020
)
}
\arguments{
\item{date_min}{a}

\item{date_max}{a}

\item{yrwk_min}{a}

\item{yrwk_max}{a}

\item{location_code}{a}

\item{granularity_geo}{a}

\item{location_reference}{a}
}
\description{
make_skeleton
}
