% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{MakeLineBrushPlot}
\alias{MakeLineBrushPlot}
\title{MakeLineBrushPlot}
\usage{
MakeLineBrushPlot(pd, x, dataVal, L2, L3, GetCols)
}
\arguments{
\item{pd}{a}

\item{x}{a}

\item{dataVal}{a}

\item{L2}{a}

\item{L3}{a}

\item{GetCols}{a}
}
\description{
MakeLineBrushPlot
}
