% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{MakeLineComparisonPlot}
\alias{MakeLineComparisonPlot}
\title{MakeLineComparisonPlot}
\usage{
MakeLineComparisonPlot(pd, x, dataVal, comparisonNames, highlightCondition,
  namesFunction = NULL, title = NULL, GetCols)
}
\arguments{
\item{pd}{a}

\item{x}{a}

\item{dataVal}{a}

\item{comparisonNames}{a}

\item{highlightCondition}{a}

\item{namesFunction}{a}

\item{title}{a}

\item{GetCols}{a}
}
\description{
MakeLineComparisonPlot
}
