% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{MakeLineExcessPlot}
\alias{MakeLineExcessPlot}
\title{MakeLineComparisonPlot}
\usage{
MakeLineExcessPlot(pd, x, dataVal, dataZ, dataCIL = NULL, dataCIU = NULL,
  allPoints = TRUE, title = NULL, pointShift = 0, xShift = 0,
  weekNumbers = FALSE, step = FALSE, GetCols)
}
\arguments{
\item{pd}{a}

\item{x}{a}

\item{dataVal}{a}

\item{dataZ}{a}

\item{dataCIL}{a}

\item{dataCIU}{a}

\item{allPoints}{a}

\item{title}{a}

\item{pointShift}{a}

\item{xShift}{a}

\item{weekNumbers}{a}

\item{step}{a}

\item{GetCols}{a}
}
\description{
MakeLineComparisonPlot
}
