% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{MakeLineThresholdPlot}
\alias{MakeLineThresholdPlot}
\title{MakeLineThresholdPlot}
\usage{
MakeLineThresholdPlot(pd, x, dataVal, dataCIL = NULL, dataCIU = NULL, L1,
  L2, L3, L4, allPoints = TRUE, title = NULL, pointShift = 0,
  xShift = 0, weekNumbers = FALSE, step = FALSE, GetCols)
}
\arguments{
\item{pd}{a}

\item{x}{a}

\item{dataVal}{a}

\item{dataCIL}{a}

\item{dataCIU}{a}

\item{L1}{a}

\item{L2}{a}

\item{L3}{a}

\item{L4}{a}

\item{allPoints}{a}

\item{title}{a}

\item{pointShift}{a}

\item{xShift}{a}

\item{weekNumbers}{a}

\item{step}{a}

\item{GetCols}{a}
}
\description{
MakeLineThresholdPlot
}
