% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msisannualreport_document.R
\name{msisannualreport_document}
\alias{msisannualreport_document}
\title{sykdompuls_mage_document}
\usage{
msisannualreport_document(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
sykdompuls_mage_document
}
