% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msisannualreport_document.R
\name{msisannualreport_resources_copy}
\alias{msisannualreport_resources_copy}
\title{Copies all necessary sykdompuls_mage_document and sykdompuls_luft_document resources}
\usage{
msisannualreport_resources_copy(output_dir)
}
\arguments{
\item{output_dir}{Folder to copy resources to}
}
\description{
Copies all necessary sykdompuls_mage_document and sykdompuls_luft_document resources
}
