% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msisannualreport_document.R
\name{msisannualreport_resources_remove}
\alias{msisannualreport_resources_remove}
\title{Removes all necessary sykdompuls_mage_document and sykdompuls_luft_document resources}
\usage{
msisannualreport_resources_remove(output_dir)
}
\arguments{
\item{output_dir}{Folder to remove resources from}
}
\description{
Removes all necessary sykdompuls_mage_document and sykdompuls_luft_document resources
}
