% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_noispiah.R
\name{noispiah_resources_copy}
\alias{noispiah_resources_copy}
\title{Copies all necessary noispiah_document resources}
\usage{
noispiah_resources_copy(output_dir)
}
\arguments{
\item{output_dir}{Folder to copy resources to}
}
\description{
Copies all necessary noispiah_document resources
}
