% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noispiah_document.R
\name{noispiah_resources_remove}
\alias{noispiah_resources_remove}
\title{Removes all necessary noispiah_document resources}
\usage{
noispiah_resources_remove(output_dir)
}
\arguments{
\item{output_dir}{Folder to remove resources from}
}
\description{
Removes all necessary noispiah_document resources
}
