% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_normomo.R
\name{normomo_resources_copy}
\alias{normomo_resources_copy}
\title{Copies all necessary normomo_document resources}
\usage{
normomo_resources_copy(output_dir)
}
\arguments{
\item{output_dir}{Folder to copy resources to}
}
\description{
Copies all necessary normomo_document resources
}
