% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_normomo.R
\name{normomo_resources_remove}
\alias{normomo_resources_remove}
\title{Removes all necessary normomo_document resources}
\usage{
normomo_resources_remove(output_dir)
}
\arguments{
\item{output_dir}{Folder to remove resources from}
}
\description{
Removes all necessary normomo_document resources
}
