% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_Documentation.R
\docType{data}
\name{PK_HT}
\alias{PK_HT}
\title{Measured flow/gauge height data}
\format{
A dataframe with 21550 rows and 2 variables:
\describe{
  \item{Date}{Date of the format, YYYY/MM/DD}
  \item{Height}{Measured flow/guage height}
 ...
}
}
\source{
Mekong River Commission
}
\usage{
PK_HT
}
\description{
A dataset containing the daily measured height of the Tonle Sap Lake at Prek Kdam from 1 January 1960
to 31 December 2018
}
\keyword{datasets}
