% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderExternally.R
\name{RenderExternally}
\alias{RenderExternally}
\title{RenderExternally}
\usage{
RenderExternally(input, output_file, output_dir, params = "x=1")
}
\arguments{
\item{input}{a}

\item{output_file}{a}

\item{output_dir}{a}

\item{params}{a}
}
\description{
RenderExternally
}
