% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotime.R
\name{start_of_season}
\alias{start_of_season}
\title{start_of_season}
\usage{
start_of_season(yrwk, start_week = 30)
}
\arguments{
\item{yrwk}{a}

\item{start_week}{the start week of the season}
}
\description{
start_of_season
}
