% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{stat_stepribbon}
\alias{stat_stepribbon}
\title{stat_stepribbon}
\usage{
stat_stepribbon(mapping = NULL, data = NULL, geom = "ribbon",
  position = "identity", inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{a}

\item{data}{a}

\item{geom}{a}

\item{position}{a}

\item{inherit.aes}{a}

\item{...}{a}
}
\description{
stat_stepribbon
}
