% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_stats.R
\name{Summary_stats}
\alias{Summary_stats}
\title{Summary statistics}
\usage{
Summary_stats(
  x,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.99,
  .drop = TRUE
)
}
\arguments{
\item{x}{input data}

\item{measurevar}{variable to be measured}

\item{groupvars}{list of sample sites}

\item{na.rm}{na values}

\item{conf.interval}{confidence interval}

\item{.drop}{something which is true}
}
\value{
Summary statistics
}
\description{
Calculates mean, standard deviation and 95% confidence intervals
}
\note{
this function is not directly called, it is required in calculating monthly threshold values e.g. TSS
}
\references{
http://www.cookbook-r.com/Manipulating_data/Summarizing_data/
}
\author{
Winston Chang \email{winston@stdout.org}
}
