% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_sykdompulsepdf.R
\name{sykdompulspdf_document}
\alias{sykdompulspdf_document}
\title{sykdompulspdf_document}
\usage{
sykdompulspdf_document(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
sykdompulspdf_document
}
