% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_sykdompulsepdf.R
\name{sykdompulspdf_resources_copy}
\alias{sykdompulspdf_resources_copy}
\title{Copies all necessary sykdompulspdf_document resources}
\usage{
sykdompulspdf_resources_copy(output_dir)
}
\arguments{
\item{output_dir}{Folder to copy resources to}
}
\description{
Copies all necessary sykdompulspdf_document resources
}
