% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_sykdompulsepdf.R
\name{sykdompulspdf_resources_remove}
\alias{sykdompulspdf_resources_remove}
\title{Removes all necessary sykdompuls_mage_document and sykdompuls_luft_document resources}
\usage{
sykdompulspdf_resources_remove(output_dir)
}
\arguments{
\item{output_dir}{Folder to remove resources from}
}
\description{
Removes all necessary sykdompuls_mage_document and sykdompuls_luft_document resources
}
