% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_environments.r
\docType{data}
\name{SYSCALLS}
\alias{SYSCALLS}
\title{Storing the last syscalls}
\format{An object of class \code{environment} of length 0.}
\usage{
SYSCALLS
}
\description{
Storing the last syscalls
}
\keyword{datasets}
