% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{ThemeShiny}
\alias{ThemeShiny}
\title{ThemeShiny}
\usage{
ThemeShiny(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base size for the theme}

\item{base_family}{Base family for the theme}
}
\description{
ThemeShiny
}
