% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_environments.r
\docType{data}
\name{VARS}
\alias{VARS}
\title{Environment to store data.tables
test}
\format{An object of class \code{environment} of length 1.}
\usage{
VARS
}
\description{
Environment to store data.tables
test
}
\keyword{datasets}
