% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoyearweek.R
\name{vec_arith.fhi_isoyearweek}
\alias{vec_arith.fhi_isoyearweek}
\title{artihmetic}
\usage{
vec_arith.fhi_isoyearweek(op, x, y, ...)
}
\arguments{
\item{op}{x}

\item{x}{x}

\item{y}{x}

\item{...}{x}
}
\description{
artihmetic
}
