% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_dir.R
\name{with_dir}
\alias{with_dir}
\title{with_dir
Taken from https://stackoverflow.com/questions/26825000/in-r-do-an-operation-temporarily-using-a-setting-such-as-working-directory}
\usage{
with_dir(dir, expr)
}
\arguments{
\item{dir}{Working directory}

\item{expr}{Expression}
}
\description{
with_dir
Taken from https://stackoverflow.com/questions/26825000/in-r-do-an-operation-temporarily-using-a-setting-such-as-working-directory
}
