% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_Documentation.R
\docType{data}
\name{Wq3S_Cam}
\alias{Wq3S_Cam}
\title{Water quality data}
\format{
A dataframe with 406 rows and 26 variables:
\describe{
  \item{Name}{Site name}
  \item{Year}{Calendar year of measurement}
  \item{Month_1}{calendar month of measurement as a number}
  \item{Month}{Calendar month of measurement as an abbreviation}
  \item{TEMP_C}{Water temperature in degrees centigrade}
  \item{pH}{pH}
  \item{TSS_mgL}{Total Suspended Solids in mg/L}
  \item{COND_mSm}{Conductivity, microSeimens/m}
  \item{Ca_meqL}{Calcium, mEq/L}
  \item{Mg_meqL}{Magnesium, mEq/L}
  \item{Na_meqL}{Sodium, mEq/L}
  \item{K_meqL}{Potassium, mEq/L}
  \item{ALK_meqL}{Alkalinity, mEq/L}
  \item{Cl_meqL}{Chloride, mEq/L}
  \item{SO4_meqL}{Sulphate, mEq/L}
  \item{NO32_mgL}{Total Nitrite and Nitrate, mg/L}
  \item{NH4N_mgL}{Ammonium, mg/L}
  \item{TOTN_mgL}{Total Nitrogen, mg/L}
  \item{TOTP_mgL}{Total Phosphorus, mg/L}
  \item{DO_mgL}{Dissolved Oxygen, mg/L}
  \item{CODMN_mgL}{Chemical Oxygen Demand, mg/L}
  \item{Ca_Mg}{Calcium/Magnesium, ratio}
  \item{Na_Cl}{Sodium/Chloride, ratio}
  \item{Na_K}{Sodium/Potassium, ratio}
  \item{Ca_SO4}{Calcium/Sulphate, ratio}
  \item{NH3_mgL}{Ammonia, mg/L}
 ...
}
}
\source{
Mekong River Commission
}
\usage{
Wq3S_Cam
}
\description{
A dataset containing water quality data collected from sites in Cambodia in the 3S river basin from October 2004
to December 2014
}
\keyword{datasets}
