% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardEmail}
\alias{DashboardEmail}
\title{Sends out mass emails}
\usage{
DashboardEmail(emailsFromExcel = NULL, emailSubject, emailText,
  emailAttachFiles = NULL, emailFooter = TRUE, BCC = TRUE,
  emailsDirect = NULL,
  XLSXLocation = PROJ$DEFAULT_EMAILS_XLSX_LOCATION,
  OAUTHLocation = PROJ$DEFAULT_EMAILS_OAUTH_LOCATION)
}
\arguments{
\item{emailsFromExcel}{A column of the excel file that holds emails}

\item{emailSubject}{Subject of the email}

\item{emailText}{Text of the email}

\item{emailAttachFiles}{Files to be attached}

\item{emailFooter}{Footer of the email}

\item{BCC}{T/F if the recipiends should be BCC'd or not}

\item{emailsDirect}{Location of the excel sheet that holds email lists}

\item{XLSXLocation}{Location of xlsx file containing email lists}

\item{OAUTHLocation}{Location of the `.httr-oauth` file for use in `gmailr`}
}
\description{
Sends out mass emails
}
