% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardEmail}
\alias{DashboardEmail}
\title{Sends out mass emails}
\usage{
DashboardEmail(emailsFromExcel = NULL, emailSubject, emailText,
  emailAttachFiles = NULL, emailFooter = TRUE, BCC = TRUE,
  emailsDirect = NULL, XLSXLocation = PROJ$DEFAULT_EMAILS_XLSX_LOCATION,
  OAUTHLocation = PROJ$DEFAULT_EMAILS_OAUTH_LOCATION)
}
\arguments{
\item{emailsFromExcel}{a}

\item{emailSubject}{a}

\item{emailText}{a}

\item{emailAttachFiles}{a}

\item{emailFooter}{a}

\item{BCC}{a}

\item{emailsDirect}{a}

\item{XLSXLocation}{a}

\item{OAUTHLocation}{a}
}
\description{
Sends out mass emails
}
