% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardInitialiseOpinionated}
\alias{DashboardInitialiseOpinionated}
\title{DashboardInitialiseOpinionated}
\usage{
DashboardInitialiseOpinionated(NAME, DEV_IF_RSTUDIO = TRUE, SILENT = FALSE)
}
\arguments{
\item{NAME}{a}

\item{DEV_IF_RSTUDIO}{a}

\item{SILENT}{a}
}
\description{
DashboardInitialiseOpinionated
}
