% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardInitialiseOpinionated}
\alias{DashboardInitialiseOpinionated}
\title{DashboardInitialiseOpinionated}
\usage{
DashboardInitialiseOpinionated(NAME, TEST_IF_RSTUDIO = TRUE)
}
\arguments{
\item{NAME}{a}

\item{TEST_IF_RSTUDIO}{a}
}
\description{
DashboardInitialiseOpinionated
}
