% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoyearweek.R
\name{isoyearweek}
\alias{isoyearweek}
\title{`isoyearweek` vector}
\usage{
isoyearweek(x = lubridate::today())
}
\arguments{
\item{x}{A numeric vector}
}
\value{
character
}
\description{
`isoyearweek` vector
}
\examples{
isoyearweek_c("2020-01-01")
}
