% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderExternally.R
\name{RenderExternally}
\alias{RenderExternally}
\title{Render a markdown file in an external R process}
\usage{
RenderExternally(input, output_file, output_dir, params = "x=1")
}
\arguments{
\item{input}{Rmd file to be rendered}

\item{output_file}{Output file}

\item{output_dir}{Output directory}

\item{params}{Any params that need to be passed to the Rmd file}
}
\description{
Render a markdown file in an external R process
}
