% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_sykdompulsepdf.R
\name{sykdompuls_mage_document}
\alias{sykdompuls_mage_document}
\title{sykdompuls_mage_document}
\usage{
sykdompuls_mage_document(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
sykdompuls_mage_document
}
